gui tabs sparc2 align: pick cleverly the detector for the fiber align
We can't just hard code to the spectrometer, as the fiber might actually be used for
other components.
=> Use .affects of the fiber aligner to know which component is to be used.

For 'safety' and compatibility with previous version, we first check for a set of known
most likely good detectors, and then just pick the first detector in the affects (so
modifiable via the microscope file).